pyproject.toml
[tool.poetry]
name = "crypto"
version = "0.1.0"
description = ""
authors = ["CS-Aditya-Rawat <adityainferno4812@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pybit = "^5.6.2"
binance-connector = "^3.6.0"
binance-futures-connector = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
